Refactor bitstream loading
Refactor bitstream loading into common code and transport-specific code.
