+++ sw/host/opentitanlib/src/transport/common/mod.rs
@@ -2,4 +2,5 @@
 // Licensed under the Apache License, Version 2.0, see LICENSE for details.
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod fpga;
 pub mod uart;
