+++ sw/host/opentitanlib/src/transport/cw310/mod.rs
@@ -10,17 +10,16 @@ use std::cell::RefCell;
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
 use std::rc::Rc;
-use std::time::Duration;
 
 use crate::io::gpio::GpioPin;
 use crate::io::spi::Target;
 use crate::io::uart::{Uart, UartError};
+use crate::transport::common::fpga::{ClearBitstream, FpgaProgram};
 use crate::transport::common::uart::SerialPortUart;
 use crate::transport::{
     Capabilities, Capability, Transport, TransportError, TransportInterfaceType,
 };
 use crate::util::parse_int::ParseInt;
-use crate::util::rom_detect::{RomDetect, RomKind};
 
 pub mod gpio;
 pub mod spi;
@@ -185,34 +184,15 @@ impl Transport for CW310 {
             // Open the console UART.  We do this first so we get the receiver
             // started and the uart buffering data for us.
             let uart = self.uart("0")?;
-            if fpga_program.bitstream.starts_with(b"__skip__") {
+            let reset_pin = self.gpio_pin(Self::PIN_SRST)?;
+            if fpga_program.skip() {
                 log::info!("Skip loading the __skip__ bitstream.");
                 return Ok(None);
             }
-            if let Some(rom_kind) = &fpga_program.rom_kind {
-                let mut rd = RomDetect::new(
-                    *rom_kind,
-                    &fpga_program.bitstream,
-                    Some(fpga_program.rom_timeout),
-                )?;
-
-                // Send a reset pulse so the ROM will print the FPGA version.
-                let reset_pin = self.gpio_pin(Self::PIN_SRST)?;
-                // Reset is active low, sleep, then drive high.
-                reset_pin.write(false)?;
-                std::thread::sleep(fpga_program.rom_reset_pulse);
-                // Also clear the UART RX buffer for improved robustness.
-                uart.clear_rx_buffer()?;
-                reset_pin.write(true)?;
-
-                // Now read the uart until the ROM prints it's version.
-                if rd.detect(&*uart)? {
-                    log::info!("Already running the correct bitstream.  Skip loading bitstream.");
-                    // If we're already running the right ROM+bitstream,
-                    // then we can skip bootstrap.
-                    return Ok(None);
-                }
+            if fpga_program.check_correct_version(&*uart, &*reset_pin)? {
+                return Ok(None);
             }
+
             // Program the FPGA bitstream.
             log::info!("Programming the FPGA bitstream.");
             let usb = self.device.borrow();
@@ -248,25 +228,7 @@ impl Transport for CW310 {
 }
 
 /// Command for Transport::dispatch().
-pub struct FpgaProgram<'a> {
-    /// The bitstream content to load into the FPGA.
-    pub bitstream: Vec<u8>,
-    /// What type of ROM to expect.
-    pub rom_kind: Option<RomKind>,
-    /// How long of a reset pulse to send to the device.
-    pub rom_reset_pulse: Duration,
-    /// How long to wait for the ROM to print its type and version.
-    pub rom_timeout: Duration,
-    /// A progress function to provide user feedback.
-    /// Will be called with the address and length of each chunk sent to the target device.
-    pub progress: Option<Box<dyn Fn(u32, u32) + 'a>>,
-}
-
-/// Command for Transport::dispatch().
 pub struct SetPll {}
 
 /// Command for Transport::dispatch(). Resets the CW310's SAM3X chip.
 pub struct ResetSam3x {}
-
-/// Command for Transport::dispatch().
-pub struct ClearBitstream {}
